% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{add_covariate}
\alias{add_covariate}
\title{Add the covariate level (e.g., habitat:forest) to the table}
\usage{
add_covariate(mcmc_summary)
}
\arguments{
\item{mcmc_summary}{data.frame output from \code{\link{summarize_mcmc}}}
}
\value{
data.frame
}
\description{
Add the covariate level to the table
}
