% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{add_model}
\alias{add_model}
\title{Add the model type (observation, detection) to the table}
\usage{
add_model(mcmc_summary)
}
\arguments{
\item{mcmc_summary}{data.frame output from \code{\link{summarize_mcmc}}}
}
\value{
data.frame
}
\description{
Add the model type (observation, detection) to the table
}
