% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tables.R
\name{add_parm}
\alias{add_parm}
\title{Add a latex style parameter name to the table}
\usage{
add_parm(mcmc_summary)
}
\arguments{
\item{mcmc_summary}{data.frame output from \code{\link{summarize_mcmc}}}
}
\value{
data.frame
}
\description{
Add a latex style parameter name to the table
}
