% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{add_species}
\alias{add_species}
\title{Add the species to the table}
\usage{
add_species(mcmc_summary)
}
\arguments{
\item{mcmc_summary}{data.frame output from \code{\link{summarize_mcmc}}}
}
\value{
data.frame
}
\description{
Add the species to the table
}
