% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{beautify_df}
\alias{beautify_df}
\title{Make MCMC output intelligible}
\usage{
beautify_df(sample_sum)
}
\arguments{
\item{sample_sum}{data.frame containing summary of MCMC output}
}
\value{
data.frame
}
\description{
The JAGS output is unintelligible without referencing the JAGS model .txt
file. This will make the output more intelligible
}
