% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_mcmc_samples.R
\name{bundle_evaluation_samples}
\alias{bundle_evaluation_samples}
\title{Function to converts MCMC lists to arrays}
\usage{
bundle_evaluation_samples(model_fit)
}
\arguments{
\item{model_fit}{list. output from \code{fit_model()}}
}
\value{
array
}
\description{
This function is used to make manipulation of the mcmc samples easier.
It also only selects the first chain
}
