% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_mcmc_samples.R
\name{bundle_mcmc_samples}
\alias{bundle_mcmc_samples}
\title{Function to bundles mcmc samples into a list}
\usage{
bundle_mcmc_samples(model_fit)
}
\arguments{
\item{model_fit}{list. output from \code{fit_model()}}
}
\value{
array
}
\description{
This function is used to make manipulation of the mcmc samples easier. The
default output from \code{rjags::jags.samples} is less predictable than
a standard array.
}
\details{
This function also concatenates the MCMC chains into one array.
}
