% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_mcmc_samples.R
\name{bundle_pred_samples}
\alias{bundle_pred_samples}
\title{Function to bundles mcmc samples into a list}
\usage{
bundle_pred_samples(model_fit)
}
\arguments{
\item{model_fit}{list. output from \code{fit_model()}}
}
\value{
array
}
\description{
This function is used to make manipulation of the mcmc samples easier.
}
