% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_waic.R
\name{calc_elppd}
\alias{calc_elppd}
\title{Calculate elppd}
\usage{
calc_elppd(host_likelihood, cowbird_likelihood)
}
\arguments{
\item{host_likelihood}{three dimensional array of likelihood estimates
from \code{\link{calc_likelihood_host}}}

\item{cowbird_likelihood}{matrix containing likelihood estimates
from \code{\link{calc_likelihood_cowbird}}}
}
\value{
numeric. estimate of the pointwise expected log predictive
posterior density
}
\description{
Calculate the pointwise expected log predictive posterior density for
the model
}
\seealso{
See Appendix S1 in \href{https://esajournals.onlinelibrary.wiley.com/doi/abs/10.1890/15-1471.1}{Broms, Hooten, and Fitzpatrick 2016}
for information about the formula
}
