% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{calc_likelihood_cowbird}
\alias{calc_likelihood_cowbird}
\title{Function to calculate the likelihood of the cowbird model}
\usage{
calc_likelihood_cowbird(Y, theta_samples, psi_samples)
}
\arguments{
\item{Y}{matrix of host detection data, number of sites by number of visits
(J x K)}

\item{theta_samples}{matrix of MCMC samples for theta parameters,
number of sites by by number of visits by number of iterations (J x K x S)}

\item{psi_samples}{matrix of MCMC samples for eta parameters,
number of sites by number of iterations}
}
\value{
matrix containing the likelihood of the
detection data given parameter values at each MCMC iteration
}
\description{
Calculate the likelihood of observations at each site given MCMC samples
of parameter estimates.
}
