% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_models.R
\name{calc_likelihood_host}
\alias{calc_likelihood_host}
\title{Function to calculate the  likelihood of the host model}
\usage{
calc_likelihood_host(X, eta_samples, zeta_samples)
}
\arguments{
\item{X}{three dimensional array of host detection data, number of sites by
number of hosts by number of visits (J x N x K)}

\item{eta_samples}{three dimensional array of MCMC samples for eta parameters,
number of sites by number of hosts by number of iterations (J x N x S)}

\item{zeta_samples}{four dimensional array of MCMC samples for eta parameters,
number of sites by number of hosts by number of visits by number of
iterations}
}
\value{
three dimensional array containing the  likelihood of the
detection data given parameter values at each MCMC iteration
}
\description{
Calculate the likelihood of observations at each site given MCMC samples
of parameter estimates.
}
