% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_naive_state.R
\name{calc_naive_state_host}
\alias{calc_naive_state_host}
\title{Naive occurrence state for hosts}
\usage{
calc_naive_state_host(detections)
}
\arguments{
\item{detections}{three dimensional array of detections, with dimensions of
number of sites by number of hosts by number of visits (J x N x K)}
}
\value{
logical matrix (J x N). was the host detected at the site?
}
\description{
Calculate the naive occurrence state for the hosts, i.e., determine
if a host was detected at a site
}
