% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_site_likelihood.R
\name{calc_site_likelihood}
\alias{calc_site_likelihood}
\title{Calculate the likelihood of the detections at a site}
\usage{
calc_site_likelihood(
  pr_detection,
  detection_history,
  pr_occupancy,
  detected_at_site
)
}
\arguments{
\item{pr_detection}{vector of detection probabilities, with length equal to
the number of visits}

\item{detection_history}{vector of detection history, with length equal to
the number visits}

\item{pr_occupancy}{numeric. the probability of occurrence at the site}

\item{detected_at_site}{logical. was the species detected at the site?}
}
\value{
numeric. likelihood of the detection data given the parameters
}
\description{
Calculates the likelihood of a detection history given parameter values
for occupancy and detection
}
\details{
This calculates the likelihood formula,
\deqn{
I * (\psi \prod p^y * (1 - p)^{1 - y}) + (1 - I) * ((1 - \psi) + \psi \prod (1 - p))
}
}
\seealso{
See Appendix S1 in \href{https://esajournals.onlinelibrary.wiley.com/doi/abs/10.1890/15-1471.1}{Broms, Hooten, and Fitzpatrick 2016}
for information about the formula
}
