% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_waic.R
\name{calc_waic}
\alias{calc_waic}
\title{Calculate WAIC}
\usage{
calc_waic(elppd, pdwaic)
}
\arguments{
\item{elppd}{numeric}

\item{pdwaic}{numeric}
}
\value{
numeric. estimate of WAIC
}
\description{
Calculate WAIC, a measure of model parsimony
}
\seealso{
See Appendix S1 in \href{https://esajournals.onlinelibrary.wiley.com/doi/abs/10.1890/15-1471.1}{Broms, Hooten, and Fitzpatrick 2016}
for information about the formula
}
