% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all_models.R
\name{check_all_models}
\alias{check_all_models}
\title{Check all models}
\usage{
check_all_models(fit_list, data_list, just_pvals = TRUE)
}
\arguments{
\item{fit_list}{list. Output from \code{\link{fit_all_models}}}

\item{data_list}{list. Output from \code{\link{make_data_list}}}

\item{just_pvals}{logical. Should the output only include p-values, or
should the function additionally return the predictions and observations?}
}
\value{
list
}
\description{
Perform posterior predictive checks for all models regarding the number of
sites where the parasite and host co-occur
}
