% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_models.R
\name{check_model}
\alias{check_model}
\title{Check a model}
\usage{
check_model(model_fit, data_list, just_pvals = TRUE)
}
\arguments{
\item{model_fit}{list. Output from \code{\link{fit_model}}}

\item{data_list}{list. Output from \code{\link{make_data_list}}}

\item{just_pvals}{logical. Should the output only include p-values, or
should the function additionally return the predictions and observations?}
}
\value{
if just_pvals, then a vector of p-values. Otherwise, a list.
}
\description{
Perform a posterior predictive check of the number of sites where the
parasite and host co-occur
}
