% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{convert_mcmc}
\alias{convert_mcmc}
\title{Convert output from fit_model to \code{\link[coda]{mcmc.list}}}
\usage{
convert_mcmc(model_fit, params = NULL)
}
\arguments{
\item{model_fit}{output from fit_model}

\item{params}{list of parameters to diagnose}
}
\value{
list of mcmc.list objects
}
\description{
Converts mcarrays to \code{\link[coda]{mcmc.list}}, the coda class for mcmc
diagnostics and processing
}
