% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_three.R, R/make_plots.R
\name{cooccur_probs}
\alias{cooccur_probs}
\title{Generate co-occurrence plot data.frame}
\usage{
cooccur_probs(model_fit)

cooccur_probs(model_fit)
}
\arguments{
\item{model_fit}{output from \code{\link{fit_model}}}
}
\value{
data.frame

data.frame
}
\description{
Generate the data.frame used in the co-occurrence plot

Generate the data.frame used in the co-occurrence plot
}
