% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_sample_size.R
\name{effective_sample_size}
\alias{effective_sample_size}
\title{Calculate the effective sample size}
\usage{
effective_sample_size(diag_samples)
}
\arguments{
\item{diag_samples}{list of mcmc.list objects}
}
\value{
data.frame
}
\description{
Calculates the effective sample sizes for mcmc parameters using
\code{\link[coda]{effectiveSize}}
}
