% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_models.R
\name{evaluate_all_models}
\alias{evaluate_all_models}
\title{Function to estimate the WAIC for all three models model}
\usage{
evaluate_all_models(fit_list, data_list)
}
\arguments{
\item{data_list}{list. output from \code{make_data_list} function.}

\item{model_fit}{list. output from \code{fit_all_models} function.}
}
\value{
data.frame of estimates for WAIC, elppd, and pdWAIC
}
\description{
WAIC is used to estimate the parsimony of each model.
}
