% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Function to estimate the WAIC for a model}
\usage{
evaluate_model(model_fit, data_list)
}
\arguments{
\item{model_fit}{list. output from \code{fit_model()} function.}

\item{data_list}{list. output from \code{make_data_list()} function.}
}
\value{
data.frame of estimates for WAIC, elppd, and pdWAIC
}
\description{
WAIC is used to estimate the parsimony of each model.
}
