% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coda_samples.R
\name{extract_coda_samples}
\alias{extract_coda_samples}
\title{Extract MCMC samples from fit_model for use in coda.}
\usage{
extract_coda_samples(model_fit, params = NULL)
}
\arguments{
\item{model_fit}{output from fit_model}

\item{params}{list of parameters to diagnose}
}
\value{
list of mcmc.list objects
}
\description{
Converts mcarrays to \code{\link[coda]{mcmc.list}}, the coda class for mcmc
diagnostics and processing
}
