% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{gelman_rubin}
\alias{gelman_rubin}
\title{Calculate the gelman rubin statistic}
\usage{
gelman_rubin(diag_samples)
}
\arguments{
\item{diag_samples}{list of mcmc.list objects}
}
\value{
data.frame
}
\description{
Calculates the gelman rubin statistic for mcmc parameters using
\code{\link[coda]{gelman.diag}}
}
