% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_all_parameters}
\alias{get_all_parameters}
\title{Generates vector of the primary parameters}
\usage{
get_all_parameters()
}
\value{
character vector of parameter names
}
\description{
Generates a vector of names of parameters, including hyper parameters,
cowbird parameters, prediction parameters, and model evaluation parameters.
}
