% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{get_cov}
\alias{get_cov}
\title{Get the covariate level (e.g., interecept or observer B.) for a parameter}
\usage{
get_cov(sample_summary)
}
\arguments{
\item{sample_summary}{data.frame. summary of mcmc samples.}
}
\value{
character vector.
}
\description{
This function is used to build the estimate table.
}
