% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_evaluation_parameters}
\alias{get_evaluation_parameters}
\title{Generates vector of parameter needed to evaluate models}
\usage{
get_evaluation_parameters(checking_parameters = FALSE)
}
\arguments{
\item{checking_parameters}{logical. should 'X.new' and 'Y.new', the
parameters used in posterior predictive checking, be included?}
}
\value{
character vector of parameter names
}
\description{
Generates vector of parameter needed to evaluate models
}
