% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_habitat_names}
\alias{get_habitat_names}
\title{Get the names of the habitats}
\usage{
get_habitat_names()
}
\value{
vector. names of the habitats
}
\description{
Get a vector of names of the habitats
}
