% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_host_full_name}
\alias{get_host_full_name}
\title{Get the full names of the host species}
\usage{
get_host_full_name()
}
\value{
vector. names of the host species
}
\description{
Get a vector of names of the host species
}
