% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_host_names}
\alias{get_host_names}
\title{Get the names of the host species}
\usage{
get_host_names()
}
\value{
vector. names of the host species
}
\description{
Get a vector of names of the host species
}
