% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tables.R
\name{get_mcmc_ci}
\alias{get_mcmc_ci}
\title{Extract the credible interval from the MCMC summary}
\usage{
get_mcmc_ci(mcmc_summary)
}
\arguments{
\item{mcmc_summary}{list. output from \code{\link{summarize_mcmc}}}
}
\value{
data.frame
}
\description{
Extract the credible interval from the MCMC summary
}
