% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{get_mod}
\alias{get_mod}
\title{Get the model type (observation or occurrence) for a parameter}
\usage{
get_mod(sample_summary)
}
\arguments{
\item{sample_summary}{data.frame. summary of mcmc samples.}
}
\value{
character vector.
}
\description{
This function is used to build the estimate table.
}
