% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_list.R
\name{get_o}
\alias{get_o}
\title{Create matrix of visit covariates affecting the hosts}
\usage{
get_o(cowbird_data)
}
\arguments{
\item{cowbird_data}{data.frame, from \code{\link{data('cowbird_data')}}}
}
\value{
matrix of visit covariates affecting the hosts
}
\description{
Create a matrix of matrix of visit covariates for the data list
}
