% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{get_parm}
\alias{get_parm}
\title{Get the full parameter name}
\usage{
get_parm(sample_summary)
}
\arguments{
\item{sample_summary}{data.frame. summary of mcmc samples.}
}
\value{
character vector.
}
\description{
This function is used to build the estimate table.
}
