% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_ppc_parameters}
\alias{get_ppc_parameters}
\title{Generates vector of parameter needed to do posterior predictive checks}
\usage{
get_ppc_parameters()
}
\value{
character vector of parameter names
}
\description{
Generates a vector of names of parameters that are used to calculate Bayesian
p-values
}
