% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_predicition_parameters}
\alias{get_predicition_parameters}
\title{Generates vector of parameter used in model predictions}
\usage{
get_predicition_parameters()
}
\value{
character vector of parameter names
}
\description{
Generates a vector of names of parameters that are used to in the predictions
of co-occurrence, number of co-occurring sites, etc.
}
