% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_table2_parameters}
\alias{get_table2_parameters}
\title{Generates vector of parameters in the Table 2}
\usage{
get_table2_parameters()
}
\value{
character vector of parameter names
}
\description{
Generates vector of parameters in the Table 2
}
