% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_list.R
\name{get_v}
\alias{get_v}
\title{Create matrix of site covariates affecting the cowbird}
\usage{
get_v(cowbird_data)
}
\arguments{
\item{cowbird_data}{data.frame, from \code{\link{data('cowbird_data')}}}
}
\value{
matrix of site covariates affecting the hosts
}
\description{
Create a matrix of matrix of site covariates for the data list
}
