% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_list.R
\name{get_w}
\alias{get_w}
\title{Create matrix of site covariates affecting the hosts}
\usage{
get_w(cowbird_data)
}
\arguments{
\item{cowbird_data}{data.frame, from the embedded cowbird_data
dataset}
}
\value{
matrix of site covariates affecting the hosts
}
\description{
Create a matrix of matrix of site covariates for the data list
}
