% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_list.R
\name{get_y}
\alias{get_y}
\title{Create matrix containing the cowbird detections}
\usage{
get_y(cowbird_data)
}
\arguments{
\item{cowbird_data}{data.frame, from \code{data('cowbird_data')}}
}
\value{
matrix containing the cowbird detections
}
\description{
Create a matrix containing the cowbird detections for the data list
}
