% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_values.R
\name{initialize_values}
\alias{initialize_values}
\title{Initialize values for MCMC}
\usage{
initialize_values(data_list)
}
\arguments{
\item{data_list}{list made from \code{make_data_list()}, containing detection
matrices for cowbird and hosts}
}
\value{
list, suitable for use in \code{rjags::jags.model()}
}
\description{
Initializes values for MCMC, namely, the occurrence states of the parasites
and hosts.
}
