% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.R
\name{is_detected}
\alias{is_detected}
\title{Determine if the species was detected}
\usage{
is_detected(detections)
}
\arguments{
\item{detections}{numeric vector of detections}
}
\value{
logical.
}
\description{
Create a logical indicator of the naive occurrence state
}
