% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_annotation_df}
\alias{make_annotation_df}
\title{Generate data.frame used in \code{\link{plot_ppc}}}
\usage{
make_annotation_df(ppc_results)
}
\arguments{
\item{ppc_results}{output from \code{\link{check_model}}}
}
\value{
data.frame
}
\description{
Generate the data.frame used in the posterior predictive checking plot.
This data.frame is used to generate the annotations of the observed number
as well as the p-values.
}
