% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_cooccur_plot_df}
\alias{make_cooccur_plot_df}
\title{Generate co-occurrence plot data.frame}
\usage{
make_cooccur_plot_df(model_fit)
}
\arguments{
\item{model_fit}{output from \code{\link{fit_model}}}
}
\value{
data.frame
}
\description{
Generate the data.frame used in the co-occurrence plot
}
