% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_list.R
\name{make_data_list}
\alias{make_data_list}
\title{Make a list of data for easy model fitting}
\usage{
make_data_list(cowbird_data)
}
\arguments{
\item{cowbird_data}{data.frame, loaded from
\code{data(cowbird_data)}}
}
\value{
list of data
}
\description{
This makes a list that can be easily used in the \code{fit_model} function
}
