% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{make_diagnostic}
\alias{make_diagnostic}
\title{Convert output from fit_model to ggs}
\usage{
make_diagnostic(model_fit)
}
\arguments{
\item{model_fit}{output from fit_model}
}
\value{
list of ggs objects
}
\description{
Converts mcarrays to ggs, a class for easy MCMC diagnostic plots, tests,
etc.
}
