% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_five.R, R/make_plots.R
\name{make_figure_five}
\alias{make_figure_five}
\title{Generate Figure 5}
\usage{
make_figure_five(model_fit)

make_figure_five(model_fit)
}
\arguments{
\item{model_fit}{output from \code{\link{fit_model}}}
}
\value{
ggplot2 object

ggplot2 object
}
\description{
Generate plot of predicted number of sites where the host occurs without
the parasite

Generate plot of predicted number of sites where the host occurs without
the parasite
}
