% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_four.R, R/make_plots.R
\name{make_figure_four}
\alias{make_figure_four}
\title{Generate Figure 4}
\usage{
make_figure_four(model_fit)

make_figure_four(model_fit)
}
\arguments{
\item{model_fit}{output from \code{\link{fit_model}}}
}
\value{
ggplot2 object

ggplot2 object
}
\description{
Generate plot of predicted number of sites with both the host and the
parasite by host and land-use.

Generate plot of predicted number of sites with both the host and the
parasite by host and land-use.
}
