% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_three.R, R/make_plots.R
\name{make_figure_three}
\alias{make_figure_three}
\title{Generate Figure 3}
\usage{
make_figure_three(model_fit)

make_figure_three(model_fit)
}
\arguments{
\item{model_fit}{output from \code{\link{fit_model}}}
}
\value{
ggplot2 object

ggplot2 object
}
\description{
Generate plot of co-occurrence probabilities between each hosts and
the Shiny Cowbird

Generate plot of co-occurrence probabilities between each hosts and
the Shiny Cowbird
}
