% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{make_table}
\alias{make_table}
\title{Make a table}
\usage{
make_table(model_fit, params)
}
\arguments{
\item{model_fit}{list. output from \code{\link{fit_model}}}
}
\value{
data.frame
}
\description{
Summarize MCMC output and add relevant descriptions to each parameter
}
