% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{make_table_one}
\alias{make_table_one}
\title{Make table 1}
\usage{
make_table_one(model_fit)
}
\arguments{
\item{model_fit}{list. output from \code{\link{fit_model}}}
}
\value{
data.frame
}
\description{
Make table 1
}
