% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_models.R
\name{multispecies_likelihood}
\alias{multispecies_likelihood}
\title{Function to calculate the  likelihood for multispecies occupancy models}
\usage{
multispecies_likelihood(detection_data, psi_samples, p_samples)
}
\arguments{
\item{detection_data}{three dimensional array of detection data, number of
sites by number of hosts by number of visits (J x N x K)}

\item{psi_samples}{three dimensional array of MCMC samples for occupancy
probabilities, with dimensions of number of sites by number of species by
number of iterations (J x N x S)}

\item{p_samples}{four dimensional array of MCMC samples for detection
probabilities, with dimensions number of sites by number of species by
number of visits by number of iterations (J x N x K x S)}
}
\value{
three dimensional array containing the  likelihood of the
detection data at each site, species, and iteration.
}
\description{
\code{multispecies_likelihood} calculates the  likelihood of detection
data for multiple species given parameter values at each MCMC sample.
}
