% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_naive_state.R
\name{multispecies_naive_state}
\alias{multispecies_naive_state}
\title{Naive occurrence state for multiple species}
\usage{
multispecies_naive_state(detections)
}
\arguments{
\item{detections}{three dimensional array of detections, with dimensions of
number of sites by number of species by number of visits (J x N x K)}
}
\value{
logical matrix (J x N). was the species detected at the site?
}
\description{
Calculate the naive occurrence state for multiple species, i.e., determine
if a species was detected at a site
}
