% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_all_ppc}
\alias{plot_all_ppc}
\title{Generate Figure 2}
\usage{
plot_all_ppc(ppc_list)
}
\arguments{
\item{ppc_list}{output from \code{\link{check_all_models}}}
}
\value{
gridExtra grob
}
\description{
Generate plot of used in posterior predictive checking
}
