% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cooccur_probs}
\alias{plot_cooccur_probs}
\title{Generate Figure 3}
\usage{
plot_cooccur_probs(model_fit)
}
\arguments{
\item{model_fit}{output from \code{\link{fit_model}}}
}
\value{
ggplot2 object
}
\description{
Generate plot of co-occurrence probabilities between each hosts and
the Shiny Cowbird
}
