% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_just_hosts}
\alias{plot_just_hosts}
\title{Generate Figure 5}
\usage{
plot_just_hosts(model_fit)
}
\arguments{
\item{model_fit}{output from \code{\link{fit_model}}}
}
\value{
ggplot2 object
}
\description{
Generate plot of predicted number of sites where the host occurs without
the parasite
}
