% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ppc}
\alias{plot_ppc}
\title{Generate posterior predictive checking plots}
\usage{
plot_ppc(ppc_results, model = c("Model 1", "Model 2", "Model 3"))
}
\arguments{
\item{ppc_results}{output from \code{\link{posterior_predictive_check}}}

\item{model}{character. specifies which model is being checked.}
}
\value{
ggplot2 object
}
\description{
Generate subplots used to generate
}
