% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_two.R, R/make_plots.R
\name{plot_ppc_individual}
\alias{plot_ppc_individual}
\title{Generate posterior predictive checking plots}
\usage{
plot_ppc_individual(ppc_results, model = c("Model 1", "Model 2", "Model 3"))

plot_ppc_individual(ppc_results, model = c("Model 1", "Model 2", "Model 3"))
}
\arguments{
\item{ppc_results}{output from \code{\link{check_model}}}

\item{model}{character. specifies which model is being checked.}
}
\value{
ggplot2 object

ggplot2 object
}
\description{
Generate subplots used to generate

Generate subplots used to generate
}
