% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sites_with_both}
\alias{plot_sites_with_both}
\title{Generate Figure 4}
\usage{
plot_sites_with_both(model_fit)
}
\arguments{
\item{model_fit}{output from \code{\link{fit_model}}}
}
\value{
ggplot2 object
}
\description{
Generate plot of predicted number of sites with both the host and the
parasite by host and land-use.
}
