% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_models.R
\name{ppc_table}
\alias{ppc_table}
\title{Tabulate results of posterior predictive check}
\usage{
ppc_table(ppc_results)
}
\arguments{
\item{ppc_results}{output from \code{\link{check_model}}}
}
\value{
matrix of p values
}
\description{
Create a table of p-values using the results of the posterior predictive
checks
}
