% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{rename_species_mat}
\alias{rename_species_mat}
\title{Convert the numerical species column to words}
\usage{
rename_species_mat(mat_sum)
}
\arguments{
\item{mat_sum}{data.frame containing summaries of the matrix parameters}
}
\value{
data.frame
}
\description{
This makes the MCMC output more intelligible
}
