% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{rename_species_vec}
\alias{rename_species_vec}
\title{Convert the numerical species column to words}
\usage{
rename_species_vec(vec_sum)
}
\arguments{
\item{vec_sum}{data.frame containing summaries of the vector parameters}
}
\value{
data.frame
}
\description{
This makes the MCMC output more intelligible
}
