% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_five.R, R/make_plots.R
\name{save_figure_five}
\alias{save_figure_five}
\title{Save Figure 5}
\usage{
save_figure_five(just_host, plot_file)

save_figure_five(just_host, plot_file)
}
\arguments{
\item{just_host}{output from \code{\link{plot_cooccur_probs}}}

\item{plot_file}{character. specifies file to save to.}
}
\description{
Save the plot to a PDF

Save the plot to a PDF
}
