% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_four.R, R/make_plots.R
\name{save_figure_four}
\alias{save_figure_four}
\title{Save Figure 4}
\usage{
save_figure_four(sites_with_both, plot_file)

save_figure_four(sites_with_both, plot_file)
}
\arguments{
\item{sites_with_both}{output from \code{\link{plot_sites_with_both}}}

\item{plot_file}{character. specifies file to save to.}
}
\description{
Save the plot to a PDF

Save the plot to a PDF
}
