% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_three.R, R/make_plots.R
\name{save_figure_three}
\alias{save_figure_three}
\title{Save Figure 3}
\usage{
save_figure_three(occurrence, plot_file)

save_figure_three(occurrence, plot_file)
}
\arguments{
\item{occurrence}{output from \code{\link{plot_cooccur_probs}}}

\item{plot_file}{character. specifies file to save to.}
}
\description{
Save the plot to a PDF

Save the plot to a PDF
}
