% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_just_host}
\alias{save_just_host}
\title{Save Figure 5}
\usage{
save_just_host(just_host, plot_file)
}
\arguments{
\item{just_host}{output from \code{\link{plot_cooccur_probs}}}

\item{plot_file}{character. specifies file to save to.}
}
\description{
Save the plot to a PDF
}
