% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_just_hosts}
\alias{save_just_hosts}
\title{Save Figure 5}
\usage{
save_just_hosts(just_hosts, plot_file)
}
\arguments{
\item{just_hosts}{output from \code{\link{plot_cooccur_probs}}}

\item{plot_file}{character. specifies file to save to.}
}
\description{
Save the plot to a PDF
}
