% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_occurrence_plot}
\alias{save_occurrence_plot}
\title{Save Figure 2}
\usage{
save_occurrence_plot(plot, plot_file)
}
\arguments{
\item{plot}{output from \code{\link{plot_cooccur_probs}}}

\item{plot_file}{character. specifies file to save to.}
}
\description{
Save the plot to a PDF
}
