% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_ppc}
\alias{save_ppc}
\title{Save Figure 2}
\usage{
save_ppc(ppc, plot_file)
}
\arguments{
\item{ppc}{output from \code{\link{plot_ppc}}}

\item{plot_file}{character. specifies file to save to.}
}
\description{
Save the plot to a PDF
}
