% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_sites_with_both}
\alias{save_sites_with_both}
\title{Save Figure 4}
\usage{
save_sites_with_both(plot, plot_file)
}
\arguments{
\item{plot}{output from \code{\link{plot_sites_with_both}}}

\item{plot_file}{character. specifies file to save to.}
}
\description{
Save the plot to a PDF
}
