% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_four.R, R/make_plots.R
\name{sites_with_both}
\alias{sites_with_both}
\title{Make data.frame used in figure four}
\usage{
sites_with_both(model_fit)

sites_with_both(model_fit)
}
\arguments{
\item{model_fit}{output from \code{\link{fit_model}}}
}
\value{
data.frame

data.frame
}
\description{
Make data.frame used in figure four

Make data.frame used in figure four
}
