% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.R
\name{sites_with_both_observed}
\alias{sites_with_both_observed}
\title{The observed number of sites with cowbird and each host}
\usage{
sites_with_both_observed(data_list)
}
\arguments{
\item{data_list}{output from \code{\link{make_data_list}}}
}
\value{
numeric vector
}
\description{
Calculate the number of sites where cowbird and each host were observed.
This is used in posterior predictive checking.
}
