% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_models.R
\name{sites_with_both_predicted}
\alias{sites_with_both_predicted}
\title{The predicted number of sites with cowbird and each host}
\usage{
sites_with_both_predicted(model_fit)
}
\arguments{
\item{model_fit}{output from \code{\link{fit_model}}}
}
\value{
numeric array
}
\description{
Calculate the number of sites where the cowbird and each host were predicted
to co-occur. This is used in posterior predictive checking.
}
