% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{summarize_matrix_params}
\alias{summarize_matrix_params}
\title{Summarize MCMC samples for matrix parameters}
\usage{
summarize_matrix_params(matrix_samples)
}
\arguments{
\item{matrix_samples}{MCMC samples for the matrix parameters}
}
\value{
data.frame
}
\description{
The JAGS code contains parameters that are either vectors (e.g., alpha) or
matrices (e.g., gamma). The latter describe the individual covariate effects
for the hosts. The former can be covariate effects from the cowbird model,
or the hyper parameters for the host model.
}
