% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{summarize_mcmc}
\alias{summarize_mcmc}
\title{Summarize the MCMC output}
\usage{
summarize_mcmc(model_fit, parameters)
}
\arguments{
\item{model_fit}{list. output from \code{\link{fit_model}}}

\item{parameters}{character vector of parameter names to summarize}
}
\value{
data.frame
}
\description{
Use coda summary.mcmc method to summarize MCMC output
}
