% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit the cowbird-host co-occurrence models}
\usage{
fit_model(
  model = c("model_1", "model_2", "model_3"),
  data_list,
  params_to_monitor = NULL,
  n.chains = 1,
  n.adapt = 100,
  n.burn = 10,
  n.iter = 10,
  n.thin = 1,
  mcmc_params_paper = FALSE,
  fix_seed = FALSE
)
}
\arguments{
\item{model}{string indicating which model is to be fit.}

\item{data_list}{list of data from \code{make_data_list()}}

\item{params_to_monitor}{character vector indicating which parameters
should be saved from the MCMC samples}

\item{n.chains}{integer representing the number of MCMC chains to run}

\item{n.adapt}{integer representing the number of MCMC iterations to adapt
the sampler}

\item{n.burn}{integer representing the number of MCMC iterations to discard
before saving samples}

\item{n.iter}{integer representing the number of MCMC iterations to save}

\item{n.thin}{integer representing the number of MCMC iterations to thin}

\item{mcmc_params_paper}{logical. should we fit the model using the MCMC
parameters described in the paper?}

\item{fix_seed}{logical. should the seed be set for reproducible results?}
}
\value{
list, elements of class \code{rjags::mcarray}
}
\description{
Fit the cowbird-host co-occurrence models described in Patton et al. (Year
TBD) using package rjags.
}
\details{
The default MCMC parameters are set for quick model fitting. To use the MCMC
parameters described in the papers, use mcmc_params_paper = TRUE.
}
\seealso{
rjags documentation for fitting models with rjags
}
