% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_hyper_parameters}
\alias{get_hyper_parameters}
\title{Generates vector of of the hyper parameters}
\usage{
get_hyper_parameters()
}
\value{
character vector of parameter names
}
\description{
Generates vector of of the hyper parameters
}
