% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_list.R
\name{get_w}
\alias{get_w}
\title{Create matrix containing the site covariates affecting the hosts}
\usage{
get_w(cowbird_data)
}
\arguments{
\item{cowbird_data}{data.frame, from \code{data('cowbird_data')}}
}
\value{
matrix containing the site covariates affecting the hosts
}
\description{
Create a matrix containing the site covariates affecting the hosts
}
