% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_list.R
\name{get_x}
\alias{get_x}
\title{Create matrix containing the host detections}
\usage{
get_x(cowbird_data)
}
\arguments{
\item{cowbird_data}{data.frame, from \code{data('cowbird_data')}}
}
\value{
matrix containing the host detections
}
\description{
Create a matrix containing the host detections for the data list
}
