% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_list.R
\name{get_x}
\alias{get_x}
\title{Create matrix of host detections}
\usage{
get_x(cowbird_data)
}
\arguments{
\item{cowbird_data}{data.frame, from \code{\link{data('cowbird_data')}}}
}
\value{
matrix of host detections
}
\description{
Create a matrix of host detections for the data list
}
