% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_list.R
\name{get_y}
\alias{get_y}
\title{Create matrix of cowbird detections}
\usage{
get_y(cowbird_data)
}
\arguments{
\item{cowbird_data}{data.frame, from the embedded cowbird_data
dataset}
}
\value{
matrix of cowbird detections
}
\description{
Create a matrix of cowbird detections for the data list
}
