% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_five.R, R/make_plots.R
\name{just_host}
\alias{just_host}
\title{Make data.frame used in figure five}
\usage{
just_host(model_fit)

just_host(model_fit)
}
\arguments{
\item{model_fit}{output from \code{\link{fit_model}}}
}
\value{
data.frame

data.frame
}
\description{
Make data.frame used in figure five

Make data.frame used in figure five
}
