% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_two.R, R/make_plots.R
\name{make_figure_two}
\alias{make_figure_two}
\title{Generate Figure 2}
\usage{
make_figure_two(ppc_list)

make_figure_two(ppc_list)
}
\arguments{
\item{ppc_list}{output from \code{\link{check_all_models}}}
}
\value{
gridExtra grob

gridExtra grob
}
\description{
Make figure two, which shows the results of posterior predictive checking
the three models.

Generate plot of used in posterior predictive checking
}
