% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{make_table_one}
\alias{make_table_one}
\title{Create a table of parameter estimates}
\usage{
make_table_one(model_fit)
}
\arguments{
\item{model_fit}{output from \code{\link{fit_model}}}
}
\value{
data.frame
}
\description{
Creates a table of parameter estimates, e.g., Tables 1
}
