% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tables.R
\name{make_table_two}
\alias{make_table_two}
\title{Make table 2}
\usage{
make_table_two(model_fit)
}
\arguments{
\item{model_fit}{list. output from \code{\link{fit_model}}}
}
\value{
data.frame
}
\description{
Make table 2
}
