% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sites_with_both}
\alias{sites_with_both}
\title{Generate co-occurrence plot data.frame}
\usage{
sites_with_both(model_fit)
}
\arguments{
\item{model_fit}{output from \code{\link{fit_model}}}
}
\value{
data.frame
}
\description{
Generate the data.frame used in the co-occurrence plot
}
